services/version_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
sqlalchemy==2.0.34
python-jose==3.3.0
asyncpg==0.29.0
httpx==0.28.1